requirements.txt
streamlit
asyncio
dotenv
aiofiles
requests
json
urllib
azure-storage-filedatalake-aio
azure-storage-filedatalake
mimetypes
logging
pprint
os
time
re
langchain-groq
langchain
sentence-transformers
groq
typing
pinecone
yt-dlp
pydub